extract trade execution logic to separate file
